+++ src/platform_impl/linux/mod.rs
@@ -359,10 +359,10 @@ impl Window {
     }
 
     #[inline]
-    pub fn set_drag_window(&self) -> Result<(), ExternalError> {
+    pub fn begin_drag(&self) -> Result<(), ExternalError> {
         match self {
             #[cfg(feature = "x11")]
-            &Window::X(ref w) => w.set_drag_window(),
+            &Window::X(ref w) => w.begin_drag(),
             #[cfg(feature = "wayland")]
             _ => Err(ExternalError::NotSupported(NotSupportedError::new())),
         }