+++ src/platform_impl/windows/window.rs
@@ -358,7 +358,7 @@ impl Window {
     }
 
     #[inline]
-    pub fn set_drag_window(&self) -> Result<(), ExternalError> {
+    pub fn begin_drag(&self) -> Result<(), ExternalError> {
         unsafe {
             let points = {
                 let mut pos = mem::zeroed();