+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - On Web, added support for `DeviceEvent::MouseMotion` to listen for relative mouse movements.
 - Added `Window::set_drag_window`. Implemented on Windows, macOS, X11 and Wayland.
 - Added `Window::set_drag_window`. Implemented on Windows, macOS and X11.
+- Added `Window::begin_drag`. Implemented on Windows, macOS and X11.
 
 # 0.24.0 (2020-12-09)